+++ claims-data/service/src/integrationTest/java/uk/gov/justice/laa/dstew/payments/claimsdata/config/AwsTestConfig.java
@@ -1,5 +1,6 @@
 package uk.gov.justice.laa.dstew.payments.claimsdata.config;
 
+import static org.testcontainers.containers.localstack.LocalStackContainer.Service.SNS;
 import static org.testcontainers.containers.localstack.LocalStackContainer.Service.SQS;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,16 +11,17 @@ import org.testcontainers.utility.DockerImageName;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.sns.SnsClient;
 import software.amazon.awssdk.services.sqs.SqsClient;
 
 @TestConfiguration
-public class SqsTestConfig {
+public class AwsTestConfig {
 
   @Bean
   public static LocalStackContainer localStack() {
     LocalStackContainer localStack =
         new LocalStackContainer(DockerImageName.parse("localstack/localstack:3.2"))
-            .withServices(SQS);
+            .withServices(SQS, SNS);
 
     localStack.start(); // start it before the SqsClient bean is created
     return localStack;
@@ -31,6 +33,19 @@ public class SqsTestConfig {
 
     return SqsClient.builder()
         .endpointOverride(localStack.getEndpointOverride(SQS))
+        .region(Region.of(localStack.getRegion()))
+        .credentialsProvider(
+            StaticCredentialsProvider.create(
+                AwsBasicCredentials.create(localStack.getAccessKey(), localStack.getSecretKey())))
+        .build();
+  }
+
+  @Bean
+  public SnsClient snsClient(@Autowired LocalStackContainer localStack) {
+    localStack.start(); // ensure container is running
+
+    return SnsClient.builder()
+        .endpointOverride(localStack.getEndpointOverride(SNS))
         .region(Region.of(localStack.getRegion()))
         .credentialsProvider(
             StaticCredentialsProvider.create(
