+++ claims-data/service/src/integrationTest/java/uk/gov/justice/laa/dstew/payments/claimsdata/controller/AbstractIntegrationTest.java
@@ -47,8 +47,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.testcontainers.containers.PostgreSQLContainer;
-import uk.gov.justice.laa.dstew.payments.claimsdata.config.SnsTestConfig;
-import uk.gov.justice.laa.dstew.payments.claimsdata.config.SqsTestConfig;
+import uk.gov.justice.laa.dstew.payments.claimsdata.config.AwsTestConfig;
 import uk.gov.justice.laa.dstew.payments.claimsdata.entity.Assessment;
 import uk.gov.justice.laa.dstew.payments.claimsdata.entity.BulkSubmission;
 import uk.gov.justice.laa.dstew.payments.claimsdata.entity.CalculatedFeeDetail;
@@ -81,7 +80,7 @@ import uk.gov.justice.laa.dstew.payments.claimsdata.util.Uuid7;
 /** This is used to isolate the common configuration for integration testing in a single class. */
 @ActiveProfiles("test")
 @SpringBootTest
-@Import({SnsTestConfig.class})
+@Import({AwsTestConfig.class})
 @AutoConfigureMockMvc
 public abstract class AbstractIntegrationTest {
